% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate predictions}
\usage{
aggregate(predictions)
}
\arguments{
\item{predictions}{(vector of predict.rfsrc() objects) The prediction objects from \code{prediction()}.}
}
\value{
something
}
\description{
This function aggregates predictions from multiple
calls of \code{prediction()}.
}
\examples{
\dontrun{
# Example usage:
prediction <- list()
predictions[[1]] <- prediction(model = "bs23.rds", ...)
predictions[[2]] <- prediction(model = "bs24.rds", ...)
result <- aggregate(predictions)
}

}
