% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_model}
\alias{aggregate_model}
\title{Aggregate models}
\usage{
aggregate_model(models)
}
\arguments{
\item{models}{(vector of `rfsrc`` objects)}
}
\value{
composite (\code{rfsrc} object) The first rfsrc object is returned
but modified with an attribute \verb{$aggregate_rank}:
each gene is given the median rank in order of importance
according to the VIMP scores from each bootstrap.
}
\description{
This function makes a composite model from many models
coming from \code{bootstrapping()}
}
\examples{
  models = c(
    paste0(rds_dir,"/bs23.rds"),
    paste0(rds_dir,"/bs24.rds"),
    paste0(rds_dir,"/bs25.rds")
  )
  ref_model <- aggregate_model(models)
  print(ref_model$aggregate_rank)
 
}
