% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrapping}
\alias{bootstrapping}
\title{Bootstrapping}
\usage{
bootstrapping(
  input,
  output,
  core_loci,
  dependent = "food",
  starts_with = "LMO",
  seed = 23,
  bootstraps = 1,
  ncores = 1L
)
}
\arguments{
\item{input}{(character) A spreadsheet describing MLST profiles, in csv or csv.gz format.}

\item{output}{(character) Output folder for bootstrap random forest models}

\item{core_loci}{(character) A comma-separated list of core loci to help remove duplicate isolates.
These loci must be present as headers in the spreadsheet from \code{input}.}

\item{dependent}{(character, default "food") The dependent variable in \code{input}}

\item{starts_with}{(character, Default: "LMO") The prefix of all independent variables.}

\item{seed}{(numeric, default: 23) Random seed}

\item{bootstraps}{(numeric, default: 1) How many random forest bootstrap models to output}

\item{ncores}{(integer, default: 1) How many cores to use. Default:1}
}
\value{
my_filenames A list of random forest filenames
}
\description{
Bootstrapping
}
\examples{
TODO
See [inst/bootstrapRF.R]

}
