% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgMLST_funs_update.R
\name{cal_conf_attr}
\alias{cal_conf_attr}
\title{cal_conf_attr}
\usage{
cal_conf_attr(n.rep, pred, train = NULL)
}
\arguments{
\item{n.rep}{}

\item{pred}{}

\item{train}{}
}
\value{
I have no idea
}
\description{
cal_conf_attr
}
\examples{
TODO
}
