% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_query}
\alias{example_query}
\title{Example query}
\format{
A data frame with 1 rows and 4804 variables:
\describe{
\item{\code{LMO0XXX}}{double Expression at site XXX}
}
}
\usage{
example_query
}
\description{
This is an example query for use with the prediction step.
}
\keyword{datasets}
