% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Listeria_isolates}
\alias{Listeria_isolates}
\title{Listeria isolates}
\format{
A data frame with 719 rows and 4807 variables:
\describe{
\item{\code{SRR_ID}}{character Sample ID}
\item{\code{SourceSite}}{character Sample Source}
\item{\code{food}}{character Food Type}
\item{\code{LMO0XXX}{double Expression at site XXX}}
}
}
\usage{
Listeria_isolates
}
\description{
This dataset provides an example spreadsheet for creating
a random forest model.
}
\keyword{datasets}
