% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgMLST_funs_update.R
\name{plot_panel_pred_prob_ind}
\alias{plot_panel_pred_prob_ind}
\title{plot_panel_pred_prob_ind}
\usage{
plot_panel_pred_prob_ind(pred, train = NULL)
}
\arguments{
\item{pred}{}

\item{train}{}
}
\value{
list(p=p)
}
\description{
plot_panel_pred_prob_ind
}
\examples{
TODO
}
