% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Predict using a random forest model}
\usage{
prediction(model_filename, query, ncores = 1L)
}
\arguments{
\item{model_filename}{(character) The filename of the Random Forest model.}

\item{query}{(character) The filename of the query data in CSV format.
This is an MLST profiles spreadsheet.}

\item{ncores}{(integer, default: 1L) The number of cores to use for parallel processing.}
}
\value{
prediction The prediction object from \code{predict.rfsrc()}
}
\description{
This function makes predictions using a random forest model.
In many situations, you would want to make predictions from
many random forest models and aggregate them downstream.
}
\examples{
# Example usage:
model_filename <- paste0(rds_dir, "/bs23.rds")
query          <- example_query
print(model_filename)
print(example_query)
result <- prediction(
  model_filename = model_filename,
  query = example_query,
  ncores = 4)

print(result$predicted)

}
