% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rds_dir}
\alias{rds_dir}
\title{rds folder}
\format{
A file path to the rds_dir
}
\usage{
rds_dir
}
\description{
This is where RDS files are stored for the unit tests.
This includes but might not be limited to:
\itemize{
\item random forest bootstrap models
\item predictions files from \code{example_query}
}
}
\keyword{datasets}
