% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgMLST_funs_update.R
\name{sel_rep_iso}
\alias{sel_rep_iso}
\title{sel_rep_iso}
\usage{
sel_rep_iso(dat, ht, cgmlst_loci)
}
\arguments{
\item{dat}{}

\item{ht}{}

\item{cgmlst_loci}{}
}
\value{
cluster.id
}
\description{
sel_rep_iso
}
\examples{
TODO
}
