% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgMLST_funs_update.R
\name{seq.rfsrc}
\alias{seq.rfsrc}
\title{seq.rfsrc}
\usage{
\method{seq}{rfsrc}(dat, imp, my.seq, ntree)
}
\arguments{
\item{dat}{}

\item{imp}{}

\item{my.seq}{}

\item{ntree}{}
}
\value{
list(res = dat)
}
\description{
seq.rfsrc
}
\examples{
TODO
}
