% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_model}
\alias{aggregate_model}
\title{Aggregate models}
\usage{
aggregate_model(models)
}
\arguments{
\item{models}{(vector of rfsrc objects)}
}
\value{
composite (rfsrc object) The first rfsrc object is returned
but modified with an attribute $aggregate_rank:
each gene is given the median rank in order of importance
according to the VIMP scores from each bootstrap.
}
\description{
This function makes a composite model from many models
coming from \code{bootstrapping()}
}
\examples{
\dontrun{
Example usage: TODO
 }

}
