% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrapping}
\alias{bootstrapping}
\title{Bootstrap random forest analysis}
\usage{
bootstrapping(
  input,
  output,
  ncores = 1L,
  bootstrap_reps = 1L,
  loci_start_with = "LMO",
  my_seed = 23L
)
}
\arguments{
\item{input}{(character) The input file path for MLST profile data in csv or csv.gz format.}

\item{output}{(character) The output directory for random forest models.}

\item{ncores}{(integer, default: 1L) The number of cores to use for parallel processing.}

\item{bootstrap_reps}{(integer, default: 1L) The number of bootstrap replicates.}

\item{loci_start_with}{(character, default: "LMO") The prefix for loci names.}

\item{my_seed}{(integer, default: 23L) The seed for reproducibility.}
}
\value{
my_filenames A list of random forest filenames
}
\description{
This function creates one or more random forest models from a spreadsheet of MLST profiles.
In many situations, you would want to make
many random forest models and make predictions from
each one downstream.
}
\examples{
\dontrun{
# Example usage:
model_filenames <- bootstrapping(input = "tests/testthat/isolates_original_plus_new_dec_1_2021.csv.gz", output = "results/",
                         ncores = 4, bootstrap_reps = 100)
}

}
