% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_query}
\alias{example_query}
\title{Example query}
\format{
A file path to example_query.csv
}
\usage{
example_query
}
\description{
This is an example query for use with the prediction step.
}
\keyword{datasets}
