% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Listeria_isolates}
\alias{Listeria_isolates}
\title{Listeria isolates}
\format{
A file path to Listeria_isolates.csv.gz
}
\usage{
Listeria_isolates
}
\description{
This dataset provides an example spreadsheet for creating
a random forest model.
}
\keyword{datasets}
