% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Predict using a random forest model}
\usage{
prediction(model_filename, query, ncores = 1L)
}
\arguments{
\item{model_filename}{(character) The filename of the Random Forest model.}

\item{query}{(character) The filename of the query data in CSV format.
This is an MLST profiles spreadsheet.}

\item{ncores}{(integer, default: 1L) The number of cores to use for parallel processing.}
}
\value{
prediction The prediction object from \code{predict.rfsrc()}
}
\description{
This function makes predictions using a random forest model.
In many situations, you would want to make predictions from
many random forest models and aggregate them downstream.
}
\examples{
\dontrun{
# Example usage:
result <- prediction(model_filename = "results/bs23.rds",
                     query = "tests/testthat/example_query.csv", ncores = 4)
}

}
