% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Prediction}
\usage{
prediction(query, model, ncores = 1L)
}
\arguments{
\item{query}{(character) A CSV file with two rows: a header and values for an MLST profile. The header should only have columns with relevant loci and not even an identifier for the genome.}

\item{model}{(character) Path to single random forest model RDS file}

\item{ncores}{(integer, default: 1L) Number of cores to use during random forest fit}
}
\value{
prediction The prediction object from \code{predict.rfsrc()}
}
\description{
Prediction
}
\examples{
See [inst/bootstrapRF.R]
}
